+++ apps/home/src/app/app.component.ts
@@ -10,10 +10,10 @@ import {
   inject,
 } from '@angular/core';
 import { RouterModule } from '@angular/router';
-import { ConnectivityService } from './shared/connectivity.service';
-import { ThemeComponent } from './shared/theme.component';
+import { ConnectivityService } from './shared/connectivity/connectivity.service';
+import { ThemeComponent } from './shared/theme/theme.component';
 import { TimeComponent } from './shared/time.component';
-import { VisibilityService } from './shared/visibility.service';
+import { VisibilityService } from './shared/visibility/visibility.service';
 import { WidgetComponent } from './shared/widget/widget.component';
 import { WidgetService } from './shared/widget/widget.service';
