+++ apps/home/src/app/shared/connectivity/connectivity.service.ts
