+++ apps/home/src/app/shared/icons/icon.pipe.ts
